Add prototype view
