+++ src/main/cpp/http/server.h
@@ -3,23 +3,15 @@
 
 #include <iostream>
 #include <unistd.h>
-#include "Poco/DateTimeFormat.h"
-#include "Poco/DateTimeFormatter.h"
 #include "Poco/Exception.h"
+#include "Poco/Net/ServerSocket.h"
 #include "Poco/Net/HTTPRequestHandler.h"
 #include "Poco/Net/HTTPRequestHandlerFactory.h"
 #include "Poco/Net/HTTPServer.h"
 #include "Poco/Net/HTTPServerParams.h"
-#include "Poco/Net/HTTPServerParams.h"
 #include "Poco/Net/HTTPServerRequest.h"
 #include "Poco/Net/HTTPServerResponse.h"
-#include "Poco/Net/ServerSocket.h"
-#include "Poco/Process.h"
 #include "Poco/ThreadPool.h"
-#include "Poco/Timestamp.h"
-#include "Poco/Util/HelpFormatter.h"
-#include "Poco/Util/Option.h"
-#include "Poco/Util/OptionSet.h"
 #include "Poco/Util/ServerApplication.h"
 
 using Poco::Net::ServerSocket;
@@ -29,25 +21,9 @@ using Poco::Net::HTTPServer;
 using Poco::Net::HTTPServerRequest;
 using Poco::Net::HTTPServerResponse;
 using Poco::Net::HTTPServerParams;
-using Poco::Timestamp;
-using Poco::DateTimeFormatter;
-using Poco::DateTimeFormat;
-using Poco::ThreadPool;
 using Poco::Util::ServerApplication;
-using Poco::Util::Application;
-using Poco::Util::Option;
-using Poco::Util::OptionSet;
-using Poco::Util::OptionCallback;
-using Poco::Util::HelpFormatter;
 
 namespace Stride {
-
-class NotFoundRequestHandler : public HTTPRequestHandler
-{
-public:
-	NotFoundRequestHandler();
-	void handleRequest(HTTPServerRequest& request, HTTPServerResponse& response);
-};
 
 class StrideRequestHandlerFactory : public HTTPRequestHandlerFactory
 {
