+++ src/main/html/index.html
@@ -1,0 +1,27 @@
+<!DOCTYPE html5>
+<html>
+    <head>
+        <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.0/jquery.min.js"></script>
+    </head>
+
+
+<body>
+    <div> DATA: <span class="data"></span> </div>
+</body>
+
+<script>
+
+$(document).ready(function () {
+    var request = new XMLHttpRequest();
+    request.open('GET', 'http://127.0.0.1:5731/API/');
+    request.responseType = 'json';
+    request.onload = function () {
+        console.log(request.response);
+        $('.data').text(request.response.message);
+    };
+    request.send();
+});
+
+
+</script>
+</html>
